feat(transformers): merge musicgen functionalities to a single backend
So we optimize space
