--- backend/python/transformers-musicgen/README.md
@@ -1,5 +1,0 @@
-# Creating a separate environment for the transformers project
-
-```
-make transformers-musicgen
-```
