--- backend/python/transformers-musicgen/requirements-cpu.txt
@@ -1,3 +1,0 @@
-transformers
-accelerate
-torch==2.4.1
