--- backend/python/transformers-musicgen/requirements-cublas11.txt
@@ -1,4 +1,0 @@
---extra-index-url https://download.pytorch.org/whl/cu118
-transformers
-accelerate
-torch==2.4.1+cu118
