--- backend/python/transformers-musicgen/requirements-hipblas.txt
@@ -1,4 +1,0 @@
---extra-index-url https://download.pytorch.org/whl/rocm6.0
-transformers
-accelerate
-torch==2.4.1+rocm6.0
