--- backend/python/transformers-musicgen/requirements-intel.txt
@@ -1,8 +1,0 @@
---extra-index-url https://pytorch-extension.intel.com/release-whl/stable/xpu/us/
-intel-extension-for-pytorch==2.3.110+xpu
-transformers
-oneccl_bind_pt==2.3.100+xpu
-accelerate
-torch==2.3.1+cxx11.abi
-optimum[openvino]
-setuptools
