--- backend/python/transformers-musicgen/run.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-source $(dirname $0)/../common/libbackend.sh
-
-startBackend $@
