--- backend/python/transformers-musicgen/test.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-set -e
-
-source $(dirname $0)/../common/libbackend.sh
-
-runUnittests
